requirements.txt
ccxt>=4.0.0
redis>=5.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
